fix: align representative workflow refactor with task 2 scope
