+++ src/Arius.E2E.Tests/Workflows/RepresentativeWorkflowCatalog.cs
@@ -6,9 +6,4 @@ internal static class RepresentativeWorkflowCatalog
 {
     internal static readonly RepresentativeWorkflowDefinition Canonical =
         new("canonical", SyntheticRepositoryProfile.Representative, 20260419, []);
-
-    public static IReadOnlyList<RepresentativeWorkflowDefinition> All { get; } =
-    [
-        Canonical,
-    ];
 }
